Issue #496  Replace or augment Tcl code with Python equivalents
Factored out getReadoutHost into rdo_utils as we're going
to use that in title setting too.
